+++ docs/source/history.rst
@@ -1,6 +1,9 @@
 Release history
 ===============
 
+Version 0.7
+  - Log the sound server name and version.
+
 Version 0.6
   - Spread out UPnP SOAP actions that start/stop a stream (issue #16).
   - Fix the ``args`` option in the [EncoderName.UDN] section of the user